Refactor/MER-275 (#137)
* reorganize spot-related entities and update package structure

* refactor spot-related imports and remove unused zone repository

* fix: update import path for SearchedSpotsList component

* refactor: rename mediaType to genericMediaType in DTOs and update references
