+++ venus/src/model/interface/spot/comment/spotCommentMediaDto.ts
@@ -3,5 +3,5 @@ import { MediaType } from "../../../enum/mediaType";
 export default interface SpotCommentMediaDto {
     id: number;
     url: string;
-    mediaType: MediaType;
+    genericMediaType: MediaType;
 }
