+++ vulcanus/src/main/java/com/merkury/vulcanus/controllers/SpotCommentController.java
@@ -3,11 +3,11 @@ package com.merkury.vulcanus.controllers;
 import com.merkury.vulcanus.exception.exceptions.CommentAccessException;
 import com.merkury.vulcanus.exception.exceptions.SpotNotFoundException;
 import com.merkury.vulcanus.exception.exceptions.UserNotFoundException;
-import com.merkury.vulcanus.features.comment.SpotCommentService;
-import com.merkury.vulcanus.model.dtos.comment.SpotCommentAddDto;
-import com.merkury.vulcanus.model.dtos.comment.SpotCommentDto;
-import com.merkury.vulcanus.model.dtos.comment.SpotCommentEditDto;
-import com.merkury.vulcanus.model.dtos.comment.SpotCommentMediaDto;
+import com.merkury.vulcanus.features.spot.SpotCommentService;
+import com.merkury.vulcanus.model.dtos.spot.comment.SpotCommentAddDto;
+import com.merkury.vulcanus.model.dtos.spot.comment.SpotCommentDto;
+import com.merkury.vulcanus.model.dtos.spot.comment.SpotCommentEditDto;
+import com.merkury.vulcanus.model.dtos.spot.comment.SpotCommentMediaDto;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
