+++ vulcanus/src/main/java/com/merkury/vulcanus/model/entities/spot/FavoriteSpot.java
@@ -1,5 +1,6 @@
-package com.merkury.vulcanus.model.entities;
+package com.merkury.vulcanus.model.entities.spot;
 
+import com.merkury.vulcanus.model.entities.UserEntity;
 import com.merkury.vulcanus.model.enums.user.dashboard.FavoriteSpotsListType;
 import jakarta.persistence.*;
 import lombok.*;
