+++ vulcanus/src/main/java/com/merkury/vulcanus/model/mappers/user/dashboard/FavoriteSpotMapper.java
@@ -2,10 +2,10 @@ package com.merkury.vulcanus.model.mappers.user.dashboard;
 
 import com.merkury.vulcanus.model.dtos.account.spots.FavoriteSpotDto;
 import com.merkury.vulcanus.model.dtos.spot.coordinates.SpotCoordinatesDto;
-import com.merkury.vulcanus.model.entities.FavoriteSpot;
-import com.merkury.vulcanus.model.entities.SpotMedia;
+import com.merkury.vulcanus.model.entities.spot.FavoriteSpot;
+import com.merkury.vulcanus.model.entities.spot.SpotMedia;
 import com.merkury.vulcanus.model.enums.GenericMediaType;
-import com.merkury.vulcanus.model.mappers.SpotTagMapper;
+import com.merkury.vulcanus.model.mappers.spot.SpotTagMapper;
 import jakarta.validation.constraints.NotNull;
 
 import java.util.Optional;
